Changes Related to Moneris Rewrite: Hosted Tokenization
bunch of backwards-compatible changes to the framework including:

* code standards
* broke up card validation method (account number, expiration date, cdc)
so the validation code can be more easily shared/customized
* no longer throw exception for add_subscriptions_support(),
add_pre_orders_support(), and
add_subscription_payment_method_change_support() due to a stale data
race condition that could not be worked around.  That exception was for
informative purposes for the developer anyhow
* added blank credit card image to use a default when type is not known
* fixed bug that broke direct framework payment gateways on the pay
page, by not including the checkout javascript when needed.  Probably
need to look into Chase Paymentech to verify we didn't break something
with the change
* New compatibility method: get_checkout_pay_page_url()
* added SV_WC_Plugin::get_home_url() convenience method for returning
the current hostname