[#61] Admin notice option to show always on settings page
Rather than assuming a message will always be shown on a plugin's
settings page, and without a "dismiss" link, this change provides a new
notice option 'always_show_on_settings' which can be set to false to
allow a message to be dismissed from the plugin settings page.

This is useful for informational/instructional messages like "Read the
docs".

ref: WC Admin Custom Order Fields