[#61] Admin notices cleanup
Couple of small fixes, and rename a lot of the implementation from
"message" to "notice".  This was inconsistent before, and the
distinction should be: a "message" is simply a bit of text, with an
associated identifier.  A "notice" displays the text in a box, with an
optional dismiss link, etc.  The user meta name of dismissed_messages is
being maintained though as it's easier than providing a migration for
every plugin.