Improved handling of tokenize-after-sale gateways
Previously a successful charge/authorize transaction followed by an
unsuccessful tokenize request in a tokenize-after-sale gateway would
mark and report the entire order as failed.  In reality the transaction
was successful, just the tokenization portion failed.

The fix is to detect this situation and mark the successful order as
"on-hold" along with an explanatory note.

related to Moneris Payment Gateway

reference ticket: http://woothemes.zendesk.com/agent/#/tickets/210543