[skyverge/wc-plugins#487] Redefine do_credit_card_transaction()
This change supports the Moneris efraud validation fix from
skyverge/wc-plugins#487 by adding an optional $response parameter, so
that the transaction can actually be performed outside of this method
(say in a child class's override) and pass in the response.

Although this is technically a backwards-incompatible change, at present
it's only the Moneris gateway itself that overrides the
do_credit_card_transaction() method.  Therefore, the only way a merchant
could get into trouble would be if they had both Moneris and another
frameworked payment gateway installed AND they updated the other payment
gateway AND the other payment gateway had this newer version of the
framework included.  seems like an acceptable level of risk