Plugin minimum required WordPress version
Framework bootstrap option to declare minimum required WordPress
version.

Note that due to the way the bootstrap class is loaded (first plugin by
directory name lexical order) we don't have a guarantee that this will
always be available, at least not until the majority of frameworked
plugins out in the wild have upgraded to this version of the framework
or greater.