Fix handling for guest pre-orders
failures that occurred during initial tokenization could result in the
initial transaction being processed immediately, rather than upon
release.

In addition, the pre-orders get_order() method failed to account for
the situation where a guest pre-order would contain a token that didn’t
exist on the gateway. Likely a legacy issue that didn’t keep pace with
development on the rest of the framework.