Workaround for GET/HEAD API requests with query params
We have a somewhat critical limitation in our current implementation of
the API request interface/classes, in that query params and body data
are not distinguished.

This means that when attempting to perform a request type that does not
support body data (GET/HEAD) any query string params were being set as
the 'body' argument in the WP http request. Which is actually fine, WP
is smart enough to handle this case, however it requires the query
string params be provided as an associative array rather than a string
as we were doing.

This commit "fixes" that bug by adding a get_params() method to the API
request interface, and using it in the API base class to build the WP
request arguments when the request method is GET or HEAD.

The limitation at this point is that methods that support both a request
body (e.g. containing an XML string) as well as query params, may use
only the request body. This is the result of us historically not
distinguishing between the query params and the request body in the API
request interface/abstract classes and should be properly fixed at some
point by making the distinction and including the URL query string
parameters in the get_request_uri() method