Set a default empty customer ID for our gateways.
This ensures WC 2.7 and our order prop compat methods do not interfere
with our customer ID feature, since as of 2.7 the
`WC_Order::get_customer_id()` method was introduced.

In the next backwards-incompatible update we’ll want to change how
we’re passing this data around, but this should take care of any cases
where the `customer_id` property is not set and ends up giving us the
user ID instead. When not set (for logged out users) WC 2.6 and before
would simply attempt `get_post_meta` via `__get()` which would return
an empty string, so that’s what I went with as the default here.