Add a method for gateways to handle capture failures
Since 5.0 we return a result array (similar to process_payment) on
capture rather than a response object. This helps us support the new
partial capture UI more easily, but there are a few gateways that are
still expecting a response object to handle some specific errors. This
mimics the standard payment behavior by calling a method that gateways
can override to inspect the response.