Rename update_token() to set_token_props()
I kept the method protected to allow subclasses to overwrite the method
if they need to set the token props in their own way.

Authorize.Net needs to set token props in its own way to make sure that
the customer_profile_id and billing data are up to date.

When the token editor in Authorize.Net saves token data it updates the
billing information from the user object, and creates a new token object
which forces the billing_hash and payment_hash to be recalculated.