Prevent jQuery conflict with Disable Admin Notices plugin
The Disable Admin Notices plugin runs callbacks attached to the
admin_notices hook and uses regular expressions to inject the Hide
notification forever link.

The regular expression assumes that only one notice is rendered per
callback. If two or more notifications (or container elements) are
rendered by the same callback, then the regular expression causes
all elements to become children of the first element.

In the notice handler case, the regular expression causes the
placeholder to become a child of one of the notifications, which
causes an exception when we try to move that notification after the
placeholder.